Update WP-Optimize – Clean, Compress, Cache.
WP-Optimize – Clean, Compress, Cache. v3.1.4

+++ plugins/wp-optimize/includes/class-updraft-smush-manager.php
@@ -50,7 +50,6 @@ class Updraft_Smush_Manager extends Updraft_Task_Manager_1_2 {
 		if (!class_exists('Updraft_Smush_Manager_Commands')) include_once('class-updraft-smush-manager-commands.php');
 		if (!class_exists('Smush_Task')) include_once('class-updraft-smush-task.php');
 		if (!class_exists('Re_Smush_It_Task')) include_once('class-updraft-resmushit-task.php');
-		if (!class_exists('Nitro_Smush_Task')) include_once('class-updraft-nitrosmush-task.php');
 		if (!class_exists('Updraft_Logger_Interface')) include_once('class-updraft-logger-interface.php');
 		if (!class_exists('Updraft_Abstract_Logger')) include_once('class-updraft-abstract-logger.php');
 		if (!class_exists('Updraft_File_Logger')) include_once('class-updraft-file-logger.php');
@@ -63,7 +62,7 @@ class Updraft_Smush_Manager extends Updraft_Task_Manager_1_2 {
 			$this->set_default_options();
 		}
 
-		$this->webservice = $this->options->get_option('compression_server', 'nitrosmush');
+		$this->webservice = $this->options->get_option('compression_server', 'resmushit');
 
 		// Ensure the saved service is valid
 		if (!in_array($this->webservice, $this->get_allowed_services())) {
@@ -88,12 +87,56 @@ class Updraft_Smush_Manager extends Updraft_Task_Manager_1_2 {
 		}
 		add_action('delete_attachment', array($this, 'unscheduled_original_file_deletion'));
 
+		add_filter('manage_media_columns', array($this, 'manage_media_columns'));
+		add_action('manage_media_custom_column', array($this, 'manage_media_custom_column'), 10, 2);
+
 		// clean backup images cron action.
 		add_action('wpo_smush_clear_backup_images', array($this, 'clear_backup_images'));
+
+		// add filter for already compressed images by EWWW Image Optimizer.
+		add_filter('wpo_get_uncompressed_images_args', array($this, 'ewww_image_optimizer_compressed_images_args'));
 
 		if (!wp_next_scheduled('wpo_smush_clear_backup_images')) {
 			wp_schedule_event(time(), 'daily', 'wpo_smush_clear_backup_images');
 		}
+	}
+
+	/**
+	 * Add custom column to Media Library.
+	 *
+	 * @param array $columns
+	 * @return mixed
+	 */
+	public function manage_media_columns($columns) {
+		$columns['wpo_smush'] = 'WP-Optimize';
+
+		return $columns;
+	}
+
+	/**
+	 * Display content in the custom column.
+	 *
+	 * @param string $column
+	 * @param int    $attachment_id
+	 */
+	public function manage_media_custom_column($column, $attachment_id) {
+		if ('wpo_smush' !== $column) return;
+
+		if (!$this->is_compressed($attachment_id)) return;
+
+		$smush_stats = get_post_meta($attachment_id, 'smush-stats', true);
+
+		$original_size = $smush_stats['original-size'];
+		$smushed_size = $smush_stats['smushed-size'];
+
+		if (0 == $original_size) {
+			$info = sprintf(__('The file was compressed to %s using WP-Optimize', 'wp-optimize'), wpo_format_filesize($smushed_size));
+		} else {
+			$saved = round((($original_size - $smushed_size) / $original_size * 100), 2);
+			$info = sprintf(__('The file was compressed from %s to %s, saving %s percent, using WP-Optimize', 'wp-optimize'), wpo_format_filesize($original_size), wpo_format_filesize($smushed_size), $saved);
+		}
+
+		echo htmlentities($info);
 	}
 
 	/**
@@ -321,13 +364,14 @@ class Updraft_Smush_Manager extends Updraft_Task_Manager_1_2 {
 	/**
 	 * Restore compressed images for selected blog.
 	 *
-	 * @param bool $restore_backup if true then restore images from backup otherwise just delete meta.
-	 * @param int  $blog_id        blog id.
-	 * @param int  $images_limit   how many images process per time.
+	 * @param bool $restore_backup 			 if true then restore images from backup otherwise just delete meta.
+	 * @param int  $blog_id        			 blog id.
+	 * @param int  $images_limit   			 how many images process per time.
+	 * @param bool $delete_only_backups_meta meta fields will deleted only for images those will restored from backup.
 	 *
 	 * @return array ['completed' => (bool), 'message' => (string), 'error' => (string)]
 	 */
-	public function bulk_restore_compressed_images($restore_backup, $blog_id = 1, $images_limit = 100) {
+	public function bulk_restore_compressed_images($restore_backup, $blog_id = 1, $images_limit = 100, $delete_only_backups_meta = false) {
 		global $wpdb;
 
 		if (is_multisite()) {
@@ -346,7 +390,7 @@ class Updraft_Smush_Manager extends Updraft_Task_Manager_1_2 {
 			$image_ids = $wpdb->get_results($wpdb->prepare("SELECT post_id FROM {$wpdb->postmeta} WHERE meta_key = 'original-file' LIMIT %d;", $images_limit), ARRAY_A);
 
 			if (!empty($image_ids)) {
-				// run restore function for each fond image.
+				// run restore function for each found image.
 				foreach ($image_ids as $image) {
 					$restore_result = $this->restore_single_image($image['post_id'], $blog_id);
 
@@ -386,14 +430,26 @@ class Updraft_Smush_Manager extends Updraft_Task_Manager_1_2 {
 		}
 
 		if ($result['completed']) {
-			if (is_multisite()) {
-				$result['message'] = sprintf(__('All the compressed images for the site %s were successfully marked as uncompressed.', 'wp-optimize'), get_site_url($blog_id));
+			if ($delete_only_backups_meta) {
+				if (is_multisite()) {
+					$result['message'] = sprintf(__('All the compressed images for the site %s were successfully restored.', 'wp-optimize'), get_site_url($blog_id));
+				} else {
+					$result['message'] = __('All the compressed images were successfully restored.', 'wp-optimize');
+				}
 			} else {
-				$result['message'] = __('All the compressed images were successfully marked as uncompressed.', 'wp-optimize');
+				if (is_multisite()) {
+					$result['message'] = sprintf(__('All the compressed images for the site %s were successfully marked as uncompressed.', 'wp-optimize'), get_site_url($blog_id));
+				} else {
+					$result['message'] = __('All the compressed images were successfully marked as uncompressed.', 'wp-optimize');
+				}
 			}
 
 			// clear all metas for smushed images after work completed.
-			$wpdb->query("DELETE FROM {$wpdb->postmeta} WHERE meta_key IN ('smush-complete', 'smush-stats', 'original-file', 'smush-info');");
+			// if $delete_only_backup_meta set to true then all meta fields was deleted in restore_single_image()
+			// and we don't need delete metas for other images.
+			if (!$delete_only_backups_meta) {
+				$wpdb->query("DELETE FROM {$wpdb->postmeta} WHERE meta_key IN ('smush-complete', 'smush-stats', 'original-file', 'smush-info');");
+			}
 		}
 
 		if (is_multisite()) {
@@ -468,7 +524,7 @@ class Updraft_Smush_Manager extends Updraft_Task_Manager_1_2 {
 	 * @param string $server - the server to test
 	 * @return bool - true if yes, false otherwise
 	 */
-	public function check_server_online($server = 'nitrosmush') {
+	public function check_server_online($server = 'resmushit') {
 		$task = $this->get_associated_task($server);
 		$online = call_user_func(array($task, 'is_server_online'));
 			
@@ -641,6 +697,7 @@ class Updraft_Smush_Manager extends Updraft_Task_Manager_1_2 {
 			'delete_image_backup_confirm'	=> __('Do you really want to delete all backup images now? This action is irreversible.', 'wp-optimize'),
 			'mark_all_images_uncompressed'	=> __('Do you really want to mark all the images as uncompressed? This action is irreversible.', 'wp-optimize'),
 			'restore_images_from_backup'	=> __('Do you want to restore the original images from the backup (where they exist?)', 'wp-optimize'),
+			'restore_all_compressed_images'	=> __('Do you really want to restore all the compressed images?', 'wp-optimize'),
 		));
 	}
 
@@ -691,7 +748,65 @@ class Updraft_Smush_Manager extends Updraft_Task_Manager_1_2 {
 			'custom'            => 100 == $options['image_quality'] || 90 == $options['image_quality'] ? false : true
 		);
 
+		$extract['compressed_by_another_plugin'] = $this->is_image_compressed_by_another_plugin($post->ID);
+
 		WP_Optimize()->include_template('admin-metabox-smush.php', false, $extract);
+	}
+
+	/**
+	 * Check if a single image compressed by another plugin.
+	 *
+	 * @param int $image_id
+	 * @return bool
+	 */
+	private function is_image_compressed_by_another_plugin($image_id) {
+		global $wpdb;
+
+		$meta = $wpdb->get_results("SELECT meta_key, meta_value FROM {$wpdb->postmeta} WHERE `post_id`={$image_id}", ARRAY_A);
+
+		if (is_array($meta)) {
+			foreach ($meta as $row) {
+				// Smush, Imagify, Compress JPEG & PNG images by TinyPNG.
+				if (in_array($row['meta_key'], array('wp-smpro-smush-data', '_imagify_optimization_level', 'tiny_compress_images'))) return true;
+				// ShortPixel Image Optimizer
+				if ('_shortpixel_status' == $row['meta_key'] && 2 <= $row['meta_key'] && 3 > $row['meta_key']) return true;
+			}
+		}
+
+		if (WP_Optimize()->get_db_info()->table_exists('ewwwio_images')) {
+			$old_show_errors = $wpdb->show_errors(false);
+			// EWWW Image Optimizer.
+			$ewww_image = $wpdb->get_col("SELECT attachment_id FROM {$wpdb->prefix}ewwwio_images WHERE `attachment_id`={$image_id} LIMIT 1");
+			if (!empty($ewww_image)) return true;
+			$wpdb->show_errors($old_show_errors);
+		}
+
+		return apply_filters('wpo_image_compressed_by_another_plugin', false);
+	}
+
+	/**
+	 * Get attachment ids for images those already compressed with EWWW Image Optimizer.
+	 * Used with filter `wpo_get_uncompressed_images_args`.
+	 *
+	 * @param array $args WP_Query arguments.
+	 * @return array
+	 */
+	public function ewww_image_optimizer_compressed_images_args($args) {
+		global $wpdb;
+
+		if (!WP_Optimize()->get_db_info()->table_exists('ewwwio_images')) return $args;
+
+		$old_show_errors = $wpdb->show_errors(false);
+		$compressed_images = $wpdb->get_col("SELECT DISTINCT(attachment_id) FROM {$wpdb->prefix}ewwwio_images");
+		$wpdb->show_errors($old_show_errors);
+
+		if (isset($args['post__not_in'])) {
+			$args['post__not_in'] = array_merge($args['post__not_in'], $compressed_images);
+		} else {
+			$args['post__not_in'] = $compressed_images;
+		}
+
+		return $args;
 	}
 
 	/**
@@ -713,14 +828,54 @@ class Updraft_Smush_Manager extends Updraft_Task_Manager_1_2 {
 			'post_status'	=> 'inherit',
 			'posts_per_page' => apply_filters('updraft_smush_posts_per_page', 1000),
 			'meta_query' => array(
-				'relation' => 'OR',
+				'relation' => 'AND',
 				array(
-					'key'	 => 'smush-complete',
-					'compare' => '!=',
-					'value'   => '1',
+					'relation' => 'OR',
+					array(
+						'key'	 => 'smush-complete',
+						'compare' => '!=',
+						'value'   => '1',
+					),
+					array(
+						'key'	 => 'smush-complete',
+						'compare' => 'NOT EXISTS',
+						'value'   => '',
+					),
 				),
+				// ShortPixel Image Optimizer plugin
 				array(
-					'key'	 => 'smush-complete',
+					'relation' => 'OR',
+					array(
+						'key'    => '_shortpixel_status',
+						'compare' => '<',
+						'value'   => '2',
+					),
+					array(
+						'key'    => '_shortpixel_status',
+						'compare' => '>=',
+						'value'   => '3',
+					),
+					array(
+						'key'	 => '_shortpixel_status',
+						'compare' => 'NOT EXISTS',
+						'value'   => '',
+					),
+				),
+				// Smush plugin
+				array(
+					'key'	 => 'wp-smpro-smush-data',
+					'compare' => 'NOT EXISTS',
+					'value'   => '',
+				),
+				// Imagify
+				array(
+					'key'	 => '_imagify_optimization_level',
+					'compare' => 'NOT EXISTS',
+					'value'   => '',
+				),
+				// Compress JPEG & PNG images by TinyPNG
+				array(
+					'key'	 => 'tiny_compress_images',
 					'compare' => 'NOT EXISTS',
 					'value'   => '',
 				),
@@ -735,6 +890,7 @@ class Updraft_Smush_Manager extends Updraft_Task_Manager_1_2 {
 				
 				switch_to_blog($site->blog_id);
 
+				$args = apply_filters('wpo_get_uncompressed_images_args', $args);
 				$images = new WP_Query($args);
 
 				foreach ($images->posts as $image) {
@@ -753,6 +909,7 @@ class Updraft_Smush_Manager extends Updraft_Task_Manager_1_2 {
 			}
 
 		} else {
+			$args = apply_filters('wpo_get_uncompressed_images_args', $args);
 			$images = new WP_Query($args);
 			foreach ($images->posts as $image) {
 				if (file_exists(get_attached_file($image->ID))) {
@@ -954,7 +1111,6 @@ class Updraft_Smush_Manager extends Updraft_Task_Manager_1_2 {
 	 */
 	public function get_allowed_services() {
 		return array(
-			'nitrosmush' => 'Nitro_Smush_Task',
 			'resmushit'  => 'Re_Smush_It_Task',
 		);
 	}
@@ -1017,6 +1173,8 @@ class Updraft_Smush_Manager extends Updraft_Task_Manager_1_2 {
 
 		// Attempt to raise limit
 		@set_time_limit(90);
+
+		$log_header = array();
 
 		// phpcs:enable
 		$log_header[] = "\n";
@@ -1237,7 +1395,8 @@ class Updraft_Smush_Manager extends Updraft_Task_Manager_1_2 {
 				$check_date = ($year_limit == $year && $month_limit == $month);
 			}
 
-			$files = glob($directory . '*-updraft-pre-smush-original.*', GLOB_BRACE);
+			// GLOB_BRACE isn't defined on some systems (Solaris, SunOS and more) > https://www.php.net/manual/en/function.glob.php
+			$files = glob($directory . '*-updraft-pre-smush-original.*', (defined('GLOB_BRACE') ? GLOB_BRACE : 0));
 
 			foreach ($files as $file) {
 				if ($check_date) {
